Overhaul the summary module.
This is now much more structured, with separate types for variables and
values, proper error handling, more traits implemented, lots of
documentation and examples, and orthogonal access to setting and
retrieving pkg_summary entries.

Lots of changes to the API, not least SummaryStream is now Summaries.

Hopefully this is a lot more idiomatic, but no doubt there are still
improvements to be made.