Overhaul and enhance plist code.
Plist entries are now parsed from a byte stream either into an
individual entry or into a populated set of entries.

Various functions allow a parsed plist to be queried.

Data is handled correctly according to whether it can be UTF-8 or
whether it needs to accept e.g. ISO-8859 for filenames.

And lots of tests and documentation!